requirements.txt
Flask==1.1.2
IMDbPY==2021.4.18
matplotlib==3.2.2
nltk==3.6.7
pandas==1.0.5
wordcloud==1.8.1
gunicorn==19.9.0
itsdangerous==2.0.1